created rock paper scissors program
The rock paper scissors program that I created was mostly made to experiment with a few different things in python, firstly I wanted to get a grasp on data validation in python and how to go about handling it, I found that it is widely accepted that use try statements is the standard so that was my approach here. I also wanted to mess around with the usage of random number generation, I found that in python it is actually easier than in many other languages I have used in the past which made that task far simpler. Lastly I wanted to get some more experience with arrays, in python they do not work super different than they do in many other languages that I have used, however I do still want to learn more about them and experiment with them as I feel I have just scratched the surface.